+++ Project02Eclipse/src/com/hemebiotech/analytics/AnalyticsCounterMain.java
@@ -1,10 +1,13 @@
 package com.hemebiotech.analytics;
 
+import com.hemebiotech.analytics.reader.ISymptomReader;
+import com.hemebiotech.analytics.reader.ReadSymptomDataFromFile;
+
 import java.util.List;
 import java.util.Map;
 
-import static com.hemebiotech.analytics.SymptomsTreeMap.countFrequencies;
-import static com.hemebiotech.analytics.WriteSymptomInFile.writeSymptomInFile;
+import static com.hemebiotech.analytics.counter.SymptomsTreeMap.countFrequencies;
+import static com.hemebiotech.analytics.writer.WriteSymptomInFile.writeSymptomInFile;
 
 // creation de sous package , ss package gestion symptom et gestion des fichiers ( creation de sous dossier et y mettre les classes correspondantes)
