--- Project02Eclipse/src/com/hemebiotech/analytics/result.out
@@ -1,23 +1,0 @@
-anxiety: 5
-arrhythmias: 3
-blindness: 1
-blurred vision: 5
-constricted pupils: 3
-cough: 6
-dialated pupils: 4
-dizziness: 5
-dry mouth: 8
-fever: 7
-headache: 3
-high blood pressure: 10
-inflamation: 7
-insomnia: 4
-low blood pressure: 4
-nausea: 5
-rapid heart rate: 1
-rash: 4
-shortness of breath: 3
-stiff neck: 4
-stomach pain: 3
-tremor: 4
-water retention: 1
